+++ electricitymap/contrib/libs/models/events.py
