+++ parsers/NI.py
@@ -7,15 +7,15 @@ from logging import Logger, getLogger
 from typing import Any, Dict, List, Optional, Union
 
 import arrow
+from parsers.lib.exceptions import ParserException
 from requests import Session
 
-from electricitymap.contrib.libs.models.datapoints import (
+from electricitymap.contrib.libs.models.events import (
     ExchangeList,
     PriceList,
     ProductionBreakdownList,
     ProductionMix,
 )
-from parsers.lib.exceptions import ParserException
 
 TIMEZONE = "America/Managua"
